Move some `simple_dom` classes to `commonMain`
